FastAPI/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pillow==9.2.0
jinja2==3.1.2
pydantic==1.10.2
keras-ocr==0.8.2
easyocr==1.6.2
pytesseract==0.3.9
paddleocr==2.6.1
opencv-python==4.6.0.66
numpy<2.0